cast the result of sizeof() - which is either 4 or 8 here - to type long
Thanks @JNechaevsky

+++ src/doom/st_stuff.c
@@ -1029,7 +1029,7 @@ ST_Responder (event_t* ev)
       M_snprintf(msg, sizeof(msg), "%s (%s) x%ld SDL%s",
                  PACKAGE_STRING,
                  BUILD_DATE,
-                 sizeof(void *) * CHAR_BIT,
+                 (long) sizeof(void *) * CHAR_BIT,
                  crispy->sdlversion);
 #undef BUILD_DATE
       plyr->message = msg;
